gh-orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gh-orchestrator"
version = "1.0.0"
description = "Automated GitHub issue orchestration with OpenCode agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Lucas Koelman"}
]
keywords = ["github", "automation", "opencode", "ci", "orchestration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
]

[project.scripts]
gh-orchestrator = "gh_orchestrator.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/agents-config"
Documentation = "https://github.com/your-org/agents-config/blob/main/README.md"
Repository = "https://github.com/your-org/agents-config"
Issues = "https://github.com/your-org/agents-config/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/gh-orchestrator/gh_orchestrator"]

[tool.pytest.ini_options]
testpaths = ["src/gh-orchestrator/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["gh_orchestrator"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
